refactor results
