+++ src/components/RandomImage/RandomImage.js
@@ -1,6 +1,7 @@
 import React from 'react';
+import './RandomImage.css';
 
-class ImagesRandom extends React.Component{
+class RandomImage extends React.Component{
     constructor(props){
         super(props);
         this.images = ['vineyard','vine','wine','grape']
@@ -11,9 +12,9 @@ class ImagesRandom extends React.Component{
     let number3 = Math.floor(Math.random() * Math.floor(this.images.length));
     let number4 = Math.floor(Math.random() * Math.floor(this.images.length));
     return (
-      <div className="ImagesRandom">
-        <div className="ImagesRandom-image">
-          <img
+      <div className="RandomImage">
+        <div className="RandomImage-image-wrapper">
+          <img className="RandomImage-image"
             src={`https://source.unsplash.com/600x400/?${this.images[number1]},${this.images[number2]},${this.images[number3]},${this.images[number4]}`}
             alt="images random"
           />
@@ -23,4 +24,4 @@ class ImagesRandom extends React.Component{
     }
 }
 
-export default ImagesRandom;
+export default RandomImage;
