+++ src/components/Result/Result.css
@@ -1,14 +1,8 @@
-
-.DisplayResults {
-    position: relative;
-    margin:0;
-}
-
-.DisplayResults-content {
+.Result-content {
     border-radius: 5px;
 }
 
-.DisplayResults-km {
+.Result-km {
     position: absolute;
     border-radius: 5px;
     margin: 10px;
@@ -20,30 +14,7 @@
     font-size: 13px;
 }
 
-.ImagesRandom {
-    margin: 10px;
-    padding: 0;
-    position: relative;
-}
-
-.ImagesRandom-image:after {
-    border-radius: 5px;
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    box-shadow: inset 0px 0px 100px black;
-}
-
-.ImagesRandom-image img {
-    border-radius: 5px;
-    width: 100%;
-    height: auto;
-}
-
-.DisplayResults-infos {
+.Result-infos {
     position: absolute;
     border-radius: 0 0 5px 5px;
     width: calc(100% - 20px);
@@ -53,7 +24,7 @@
     background: linear-gradient(to bottom, rgba(0, 0, 0, 0), rgba(0, 0, 0, 0.1), rgba(0, 0, 0, 0.3), rgba(0, 0, 0, 0.6));
 }
 
-.DisplayResults-infos-title {
+.Result-infos-title {
     color: #F7F4E9;
     font-weight: 600;
     text-shadow: 0px 0px 1px black;
@@ -63,7 +34,7 @@
     margin-bottom: 5px;
 }
 
-.DisplayResults-infos-desc {
+.Result-infos-desc {
     color: #F7F4E9;
     margin-bottom: 0;
     font-style: italic;
