--- src/components/Results.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-class Results extends React.Component{
-    render(){
-        return( 
-            <div>
-                halo
-            </div>
-        )
-    }
-}
-export default Results
