+++ src/components/Results/Results.css
@@ -1,0 +1,6 @@
+
+.Results {
+    position: relative;
+    margin:0;
+}
+
